requirements.txt
Django==2.2.7
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
django-haystack==2.8.1
django-nine==0.2.2
djangorestframework==3.10.3
elasticsearch==7.1.0
elasticsearch-django==6.4
elasticsearch-dsl==7.1.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
tqdm==4.39.0
urllib3==1.25.7
